requirements.txt
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
distlib==0.3.6
filelock==3.12.0
Flask==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
nltk==3.8.1
numpy==1.26.0
opencv-python==4.8.1.78
platformdirs==3.5.1
pypng==0.20220715.0
python-decouple==3.8
pyzbar==0.1.9
qrcode==7.4.2
regex==2023.8.8
requests==2.31.0
textblob==0.17.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.6
virtualenv==20.23.0
Werkzeug==3.0.0
